requirements.txt
aniso8601==9.0.1
attrs==21.2.0
click==8.0.3
Flask~=2.1.2
flask-restx==0.5.1
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==4.1.2
MarkupSafe==2.0.1
pyrsistent==0.18.0
pytz==2021.3
six==1.16.0
typing-extensions==3.10.0.2
Werkzeug==2.0.2
zipp==3.6.0

marshmallow~=3.18.0
PyJWT~=2.5.0